Rename some viewmodel classes and extract their binding methods into BindingImpl classes. Add loading spinner to glide
