+++ app/build.gradle
@@ -37,6 +37,7 @@ dependencies {
     implementation "androidx.appcompat:appcompat:1.2.0"
     implementation "androidx.core:core-ktx:1.3.1"
     implementation "androidx.constraintlayout:constraintlayout:2.0.1"
+    implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.1.0"
     implementation "androidx.cardview:cardview:1.0.0"
     implementation "androidx.recyclerview:recyclerview:1.1.0"
     implementation 'androidx.palette:palette-ktx:1.0.0'
