+++ app/src/main/java/uk/co/jakebreen/pokecart/ui/shop/ShopAdapter.kt
@@ -9,23 +9,23 @@ import uk.co.jakebreen.pokecart.R
 import uk.co.jakebreen.pokecart.databinding.ShopItemBinding
 
 
-class ShopAdapter internal constructor(private val viewModels: MutableList<ShopItemViewModel>,
+class ShopAdapter internal constructor(private val items: MutableList<ShopItem>,
                                        private val clickListener: ShopViewModelClickListener):
-    RecyclerView.Adapter<ShopAdapter.ShopViewHolder>() {
+    RecyclerView.Adapter<ShopAdapter.ShopItemViewHolder>() {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ShopViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ShopItemViewHolder {
         val binding = DataBindingUtil.inflate<ShopItemBinding>(LayoutInflater.from(parent.context), R.layout.shop_item, parent, false)
-        return ShopViewHolder(binding)
+        return ShopItemViewHolder(binding)
     }
 
-    override fun getItemCount() = viewModels.size
+    override fun getItemCount() = items.size
 
-    override fun onBindViewHolder(holder: ShopViewHolder, position: Int) {
-        holder.bind(viewModels.elementAt(position), clickListener)
+    override fun onBindViewHolder(holder: ShopItemViewHolder, position: Int) {
+        holder.bind(items.elementAt(position), clickListener)
     }
 
-    class ShopViewHolder(private val binding: ShopItemBinding): RecyclerView.ViewHolder(binding.root) {
-        fun bind(viewModel: ShopItemViewModel, clickListener: ShopViewModelClickListener) {
+    class ShopItemViewHolder(private val binding: ShopItemBinding): RecyclerView.ViewHolder(binding.root) {
+        fun bind(viewModel: ShopItem, clickListener: ShopViewModelClickListener) {
             binding.viewModel = viewModel
             binding.viewModelClick = clickListener
             binding.executePendingBindings()
@@ -35,14 +35,14 @@ class ShopAdapter internal constructor(private val viewModels: MutableList<ShopI
         }
     }
 
-    fun updateAll(updated: List<ShopItemViewModel>) {
+    fun updateAll(updated: List<ShopItem>) {
         val diffResult = DiffUtil.calculateDiff(
-            ShopItemViewModelDiffCallback(
-                viewModels,
+            ShopItemDiffCallback(
+                items,
                 updated
             ), false)
-        viewModels.clear()
-        viewModels.addAll(updated)
+        items.clear()
+        items.addAll(updated)
         diffResult.dispatchUpdatesTo(this)
     }
